+++ Repository/DTO/ResponseDTO/Invoice/InvoiceOwnerView.cs
@@ -12,8 +12,8 @@ namespace Repository.DTO.ResponseDTO.Invoice
         public string InvoiceNo { get; set; }
         public DateTime IssueDate { get; set; }
         public DateTime DueDate { get; set; }
-        public double SubTotal { get; set; }
-        public double GrandTotal { get; set; }
+        public decimal SubTotal { get; set; }
+        public decimal GrandTotal { get; set; }
         public string? Note { get; set; }
         public DateTime CreateDate { get; set; }
         public Guid CustomerId { get; set; }
