+++ Repository/Repositories/Interfaces/IPaymentRepository.cs
@@ -12,8 +12,8 @@ namespace Repository.Repositories.Interfaces
     {
         Task<PaymentHistory?> CreateNewPaymentForBookingFee(Guid invoiceId);
         Task<PaymentHistory?> CreateNewPaymentForRentalFee(Guid invoiceId);
-        Task<PaymentHistory?> CreateNewPaymentForFineFee(Guid invoiceId, double fine);
-        Task<PaymentHistory?> CreateNewPaymentForAdditionFee(Guid invoiceId, double fee);
+        Task<PaymentHistory?> CreateNewPaymentForFineFee(Guid invoiceId, decimal fine);
+        Task<PaymentHistory?> CreateNewPaymentForAdditionFee(Guid invoiceId, decimal fee);
         Task<PaymentHistory?> GetPaymentById(Guid paymentId);
         Task<PaymentHistory?> UpdatePaymentStatusAndMethod(Guid paymentId, string status, string method);
         Task<List<PaymentHistory>?> GetPaymentsByUserId(Guid userId);
